edit hackathon nav and start working on your team page
